Calculate summary of inundation data

This gives summary information of the inundation data returned by calc_inundation. Included are: max inundation days by water year, total inundation days per water year, date of and flood peak value (max of dayflow) by water year, first and last overtopping date by water year, last day of flooding number of overtopping events

calc_summary()

Value

data.frame A data frame of inundation summary data.

Examples

summary <- calc_inundation()
#> Rows: 354186 Columns: 9
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (6): STATION_ID, DURATION, SENSOR_TYPE, VALUE, DATA_FLAG, UNITS
#> dbl  (1): SENSOR_NUMBER
#> dttm (2): DATE TIME, OBS DATE
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 
#> Rows: 3744 Columns: 34
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (34): Year, Month, Date, AMR, BRCR, CALR, CSMR, DRCR, FCSL, GCD, MISDV, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 3653 Columns: 38
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (38): Year, Month, Date, AMR, BRCR, CALR, CCC, CCGO, CCGCT, CSMR, DRCR, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 365 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Year, Mo, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CCC, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 5113 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Y, Year, Month, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT,...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2099 Columns: 38
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (38): Year, Month, Date, AMR, BRCR, CALR, CCC, CCGO, CCGCT, CSMR, DRCR, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 365 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Year, Mo, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CCC, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 365 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Year, Mo, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CCC, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 4749 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Y, Year, Month, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT,...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 366 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Year, Mo, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CCC, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 365 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Year, Mo, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CCC, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 9861 Columns: 29
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (29): Year, Month, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CC...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 5114 Columns: 28
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (28): Year, Month, Date, SAC, YOLO, CSMR, MOKE, MISC, SJR, EAST, TOT, CC...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.